Title: 사각형 채우기 3, Time: 0ms, Memory: 0MB, Status: Passed - Codetree

+++ 260413/사각형 채우기 3/rectangle-fill-3.java
@@ -5,12 +5,12 @@ public class Main {
     public static void main(String[] args) throws IOException{
         BufferedReader bf = new BufferedReader(new InputStreamReader(System.in));
         int n = Integer.parseInt(bf.readLine());
-        int[] dp = new int[1001];
+        long[] dp = new long[1001];
 
         dp[0] = 1;
         dp[1] = 2;
         dp[2] = 7;
-        int sum = 1;
+        long sum = 1;
 
         for(int i=3;i<=n;i++) {
             dp[i] = ((dp[i-2]*3) + (dp[i-1]*2) + 2*sum)%1000000007;
